+++ README.md
@@ -41,5 +41,6 @@ Check out [this project's issues](https://github.com/JamesHawkinss/movie-web/iss
 - [ ] Add github and discord top right
 - [ ] Store watched percentage
 - [ ] Implement movie + series view
-- [ ] Add scraper stream method
+- [ ] Add provider stream method
 - [ ] Better looking error boundary
+- [ ] sort search results so they aren't sorted by provider
